add api to init student knowledge score

+++ src/main/resources/db/migration/V1.1__create_student_knowledge_ability_table.sql
@@ -2,9 +2,9 @@ CREATE TABLE `student_knowledge_ability` (
   `id` char(36) NOT NULL,
   `student_id` char(36) NOT NULL,
   `sub_content_id` char(36) NOT NULL,
-  `current_score` INT NOT NULL,
-  `target_score` INT NOT NULL,
-  `base_score` INT NOT NULL,
+  `current_score` decimal(6,2) NOT NULL,
+  `target_score` decimal(6,2) NOT NULL,
+  `base_score` decimal(6,2) NOT NULL,
   `created_at`      TIMESTAMP NULL DEFAULT CURRENT_TIMESTAMP,
   `updated_at`      TIMESTAMP NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
   PRIMARY KEY (`id`)
